commit avant suppression relation

+++ src/Controller/ArticlesController.php
@@ -16,10 +16,6 @@ class ArticlesController extends AbstractController
     #[Route('/articles', name: 'app_articles')]
     public function index(ManagerRegistry $doctrine): Response
     {
-        $user = $this->getUser();
-        if (!$user instanceof User || (!$user->hasRole('ADMIN')) && !$user->hasRole('USER')) {
-            return $this->redirectToRoute('app_login');
-        }
         $article = $doctrine->getRepository(Articles::class)->findAll();
         return $this->render('articles/articles.html.twig', [
             'article' => $article,
@@ -42,12 +38,16 @@ class ArticlesController extends AbstractController
 
         $formArticle->handleRequest($request);
         if ($formArticle->isSubmitted() && $formArticle->isValid()) {
+
+            $user = $this->getUser();
+            $article->setUser($user);
+
             $entityManager->persist($article);
             $entityManager->flush();
 
             $this->addFlash('success', "L'article a bien été ajoutée !");
 
-            return $this->redirectToRoute('app_home');
+            return $this->redirectToRoute('app_articles');
         }
 
         return $this->render('articles/form-add.html.twig', ['formArticle' => $formArticle->createView()]);
@@ -63,6 +63,10 @@ class ArticlesController extends AbstractController
     #[Route('/articles/edit/{id}', name: 'articles_edit')]
     public function update(ManagerRegistry $doctrine, int $id, Request $request): Response
     {
+        $user = $this->getUser();
+        if (!$user instanceof User || (!$user->hasRole('ADMIN')) && !$user->hasRole('USER')) {
+            return $this->redirectToRoute('app_login');
+        }
 
         $entityManager = $doctrine->getManager();
         $article = $doctrine->getRepository(Articles::class)->find($id);
@@ -73,7 +77,7 @@ class ArticlesController extends AbstractController
         $formArticle->handleRequest($request);
         if ($formArticle->isSubmitted() && $formArticle->isValid()) {
             $entityManager->flush();
-            return $this->redirectToRoute('app_home');
+            return $this->redirectToRoute('app_articles');
         }
 
         return $this->render('articles/form-edit.html.twig', ['formArticle' => $formArticle->createView()]);
@@ -82,11 +86,17 @@ class ArticlesController extends AbstractController
     #[Route('/articles/delete/{id}', name: 'articles_delete')]
     public function delete(ManagerRegistry $doctrine, $id)
     {
+        $user = $this->getUser();
+        if (!$user instanceof User || (!$user->hasRole('ADMIN')) && !$user->hasRole('USER')) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $article = $doctrine->getRepository(Articles::class)->find($id);
         $entityManager = $doctrine->getManager();
+
         $entityManager->remove($article);
         $entityManager->flush();
 
-        return $this->redirectToRoute('app_home');
+        return $this->redirectToRoute('app_articles');
     }
 }
